Fix key been overwritten
A public key must be unique for a user, even if the key type is
different. In that case, the last case was overwritten the previous one
making the tests fails.

Also, refactor tests to make sure variables are not reused from a
previous user. To do that, each user receive it's own keys.